pyproject.toml
[project]
name = "bank-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Henrique Jesus",email = "henrique.jesus128@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "bank_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
